PAI LAB TASK 12/requirements.txt
flask
sentence-transformers
faiss-cpu
numpy